+++ src/prefect/tasks/control_flow/__init__.py
@@ -1,2 +1,9 @@
-from prefect.tasks.control_flow.conditional import ifelse, switch, merge
+from prefect.tasks.control_flow.conditional import (
+    ifelse,
+    switch,
+    merge,
+    conditional,
+    any_condition,
+    all_conditions,
+)
 from prefect.tasks.control_flow.filter import FilterTask