+++ tests/test_triggers.py
@@ -38,46 +38,20 @@ def test_all_successful_with_all_success():
 
 def test_all_successful_with_all_success_or_skipped():
     # True when all successful or skipped
-    assert triggers.all_successful(generate_states(success=3, skipped=3))
+    with pytest.raises(signals.SKIP):
+        assert triggers.all_successful(generate_states(success=3, skipped=3))
 
 
 def test_all_successful_with_all_failed():
     # Fail when all fail
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.all_successful(generate_states(failed=3))
 
 
 def test_all_successful_with_some_failed():
     # Fail when some fail
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.all_successful(generate_states(failed=3, success=1))
-
-
-def test_not_all_skipped_with_all_success():
-    # True when all successful
-    assert triggers.not_all_skipped(generate_states(success=3))
-
-
-def test_not_all_skipped_with_all_skipped():
-    with pytest.raises(signals.SKIP) as exc:
-        assert triggers.not_all_skipped(generate_states(skipped=3))
-
-
-def test_not_all_skipped_with_all_success_or_skipped():
-    # True when all successful or skipped
-    assert triggers.not_all_skipped(generate_states(success=3, skipped=3))
-
-
-def test_not_all_skipped_with_all_failed():
-    # Fail when all fail
-    with pytest.raises(signals.TRIGGERFAIL):
-        triggers.not_all_skipped(generate_states(failed=3))
-
-
-def test_not_all_skipped_with_some_failed():
-    # Fail when some fail
-    with pytest.raises(signals.TRIGGERFAIL):
-        triggers.not_all_skipped(generate_states(failed=3, success=1))
 
 
 def test_all_failed_with_all_failed():
@@ -85,12 +59,12 @@ def test_all_failed_with_all_failed():
 
 
 def test_all_failed_with_some_success():
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         assert triggers.all_failed(generate_states(failed=3, success=1))
 
 
 def test_all_failed_with_some_skips():
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         assert triggers.all_failed(generate_states(failed=3, skipped=1))
 
 
@@ -104,7 +78,7 @@ def test_always_run_with_all_failed():
 
 def test_always_run_with_mixed_states():
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.always_run(generate_states(success=1, failed=1, skipped=1, retrying=1))
 
 
@@ -158,7 +132,7 @@ def test_all_finished_with_mixed_states():
 
 
 def test_all_finished_with_some_pending():
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.all_finished(generate_states(success=1, pending=1))
 
 
@@ -170,16 +144,22 @@ def test_any_successful_with_some_success_and_some_skip():
     assert triggers.any_successful(generate_states(success=3, skipped=3))
 
 
+def test_any_successful_with_all_skipped():
+    with pytest.raises(signals.SKIP):
+        assert triggers.any_successful(generate_states(skipped=1))
+
+
 def test_any_successful_with_some_failed_and_1_success():
     assert triggers.any_successful(generate_states(failed=3, success=1))
 
 
 def test_any_successful_with_some_failed_and_1_skip():
-    assert triggers.any_successful(generate_states(failed=3, skipped=1))
+    with pytest.raises(signals.SKIP):
+        assert triggers.any_successful(generate_states(failed=3, skipped=1))
 
 
 def test_any_successful_with_all_failed():
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.any_successful(generate_states(failed=3))
 
 
@@ -196,7 +176,7 @@ def test_any_failed_with_some_failed_and_1_success():
 
 
 def test_any_failed_with_all_success():
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         triggers.any_failed(generate_states(success=3))
 
 
@@ -214,7 +194,7 @@ def test_some_failed_with_no_args(states):
 
 def test_some_failed_error_msg():
     trigger = triggers.some_failed(at_least=23)
-    with pytest.raises(signals.TRIGGERFAIL, match="some_failed"):
+    with pytest.raises(signals.SKIP, match="some_failed"):
         trigger(generate_states(success=1))
 
 
@@ -222,20 +202,20 @@ def test_some_failed_with_one_arg():
     trigger = triggers.some_failed(at_least=4)
     assert trigger(generate_states(failed=4, skipped=2))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=3))
 
     trigger = triggers.some_failed(at_most=4)
     assert trigger(generate_states(failed=4, skipped=2))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=5))
 
 
 @pytest.mark.parametrize("at_least,at_most", [(1, 1), (0.2, 1), (0.01, 20)])
 def test_some_failed_with_all_success(at_least, at_most):
     trigger = triggers.some_failed(at_least=at_least, at_most=at_most)
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=3))
 
 
@@ -243,19 +223,19 @@ def test_some_failed_does_the_math():
     trigger = triggers.some_failed(at_least=0.1, at_most=2)
     assert trigger(generate_states(failed=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=1, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=3, pending=18))
 
     trigger = triggers.some_failed(at_least=2, at_most=0.1)
     assert trigger(generate_states(failed=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=1, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=3, pending=18))
 
 
@@ -264,7 +244,7 @@ def test_some_failed_is_pickleable():
     new_trigger = cloudpickle.loads(cloudpickle.dumps(trigger))
     assert new_trigger(generate_states(failed=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         new_trigger(generate_states(failed=1, pending=18))
 
 
@@ -282,7 +262,7 @@ def test_some_successful_with_no_args(states):
 
 def test_some_successful_error_msg():
     trigger = triggers.some_successful(at_least=23)
-    with pytest.raises(signals.TRIGGERFAIL, match="some_successful"):
+    with pytest.raises(signals.SKIP, match="some_successful"):
         trigger(generate_states(success=1))
 
 
@@ -290,20 +270,20 @@ def test_some_successful_with_one_arg():
     trigger = triggers.some_successful(at_least=4)
     assert trigger(generate_states(success=4, failed=2))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=3))
 
     trigger = triggers.some_successful(at_most=4)
     assert trigger(generate_states(failed=4, pending=5))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=5))
 
 
 @pytest.mark.parametrize("at_least,at_most", [(1, 1), (0.2, 1), (0.01, 20)])
 def test_some_successful_with_all_failed(at_least, at_most):
     trigger = triggers.some_successful(at_least=at_least, at_most=at_most)
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(failed=3))
 
 
@@ -311,19 +291,19 @@ def test_some_successful_does_the_math():
     trigger = triggers.some_successful(at_least=0.1, at_most=2)
     assert trigger(generate_states(success=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=1, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=3, pending=18))
 
     trigger = triggers.some_successful(at_least=2, at_most=0.1)
     assert trigger(generate_states(success=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=1, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         trigger(generate_states(success=3, pending=18))
 
 
@@ -332,7 +312,7 @@ def test_some_successful_is_pickleable():
     new_trigger = cloudpickle.loads(cloudpickle.dumps(trigger))
     assert new_trigger(generate_states(success=2, pending=18))
 
-    with pytest.raises(signals.TRIGGERFAIL):
+    with pytest.raises(signals.SKIP):
         new_trigger(generate_states(success=1, pending=18))
 
 
@@ -356,3 +336,20 @@ def test_stateful_triggers_return_true_for_empty_upstream(func):
     assert func()(set()) is True
     assert func(at_least=2, at_most=4)(set()) is True
     assert func(at_least=0.1, at_most=0.5)(set()) is True
+
+
+def test_not_all_skipped_with_some_skipped_and_failed():
+    assert triggers.not_all_skipped(generate_states(failed=1, skipped=3)) is True
+
+
+def test_not_all_skipped_with_some_skipped():
+    assert triggers.not_all_skipped(generate_states(success=1, skipped=3)) is True
+
+
+def test_not_all_skipped_with_all_successful():
+    assert triggers.not_all_skipped(generate_states(success=3)) is True
+
+
+def test_not_all_skipped_with_all_skipped():
+    with pytest.raises(signals.SKIP):
+        triggers.not_all_skipped(generate_states(skipped=3))